Add sample input and output
Prior to this change, it was difficult for someone new to try out this
script, as they needed to provide their own input data in the expected
format (currently ChatGPT).

This change adds a sample input file and outputs so that anyone can try
it out, or have a look at the plots that are generated.